Stashing changes

+++ src/app/beer_garden/api/http/handlers/vbeta/file.py
@@ -6,6 +6,7 @@ from brewtils.schema_parser import SchemaParser
 
 from beer_garden.api.http.base_handler import BaseHandler
 from beer_garden.db.mongo.models import RawFile
+from beer_garden.garden import local_garden
 
 
 class RawFileAPI(BaseHandler):
@@ -31,6 +32,9 @@ class RawFileAPI(BaseHandler):
         tags:
           - Files
         """
+        self.minimum_permission = self.READ_ONLY
+
+        self.verify_user_permission_for_object(local_garden())
         db_file = RawFile.objects.get(id=file_id)
         file = db_file.file.read()
 
@@ -59,6 +63,9 @@ class RawFileAPI(BaseHandler):
         tags:
           - Files
         """
+        self.minimum_permission = self.OPERATOR
+
+        self.verify_user_permission_for_object(local_garden())
         db_file = RawFile.objects.get(id=file_id)
         db_file.file.delete()
         db_file.save()
@@ -88,6 +95,9 @@ class RawFileListAPI(BaseHandler):
         tags:
           - Files
         """
+        self.minimum_permission = self.OPERATOR
+
+        self.verify_user_permission_for_object(local_garden())
         db_file = RawFile()
         db_file.file.put(io.BytesIO(self.request.body))
         db_file.save()
